+++ lib/Core/Executor.cpp
@@ -25,9 +25,7 @@
 #include "TimingSolver.h"
 #include "UserSearcher.h"
 
-#include "klee/ADT/KTest.h"
 #include "klee/ADT/RNG.h"
-#include "klee/Common.h"
 #include "klee/Config/Version.h"
 #include "klee/Core/Interpreter.h"
 #include "klee/Expr/ArrayExprOptimizer.h"
@@ -36,11 +34,14 @@
 #include "klee/Expr/ExprPPrinter.h"
 #include "klee/Expr/ExprSMTLIBPrinter.h"
 #include "klee/Expr/ExprUtil.h"
+#include "klee/Solver/Common.h"
+#include "klee/ADT/KTest.h"
+#include "klee/Support/OptionCategories.h"
+#include "klee/Statistics/TimerStatIncrementer.h"
 #include "klee/Module/Cell.h"
 #include "klee/Module/InstructionInfoTable.h"
 #include "klee/Module/KInstruction.h"
 #include "klee/Module/KModule.h"
-#include "klee/OptionCategories.h"
 #include "klee/Solver/SolverCmdLine.h"
 #include "klee/Solver/SolverStats.h"
 #include "klee/Support/ErrorHandling.h"
@@ -49,7 +50,6 @@
 #include "klee/Support/ModuleUtil.h"
 #include "klee/System/MemoryUsage.h"
 #include "klee/System/Time.h"
-#include "klee/TimerStatIncrementer.h"
 
 #include "llvm/ADT/SmallPtrSet.h"
 #include "llvm/ADT/StringExtras.h"
