+++ lib/Core/SeedInfo.cpp
@@ -13,10 +13,10 @@
 #include "Memory.h"
 #include "TimingSolver.h"
 
-
 #include "klee/ADT/KTest.h"
 #include "klee/Expr/Expr.h"
 #include "klee/Expr/ExprUtil.h"
+#include "klee/ADT/KTest.h"
 #include "klee/Support/ErrorHandling.h"
 
 using namespace klee;
