+++ lib/Core/StatsTracker.cpp
@@ -12,7 +12,6 @@
 #include "ExecutionState.h"
 
 #include "klee/Config/Version.h"
-
 #include "klee/Module/InstructionInfoTable.h"
 #include "klee/Module/KInstruction.h"
 #include "klee/Module/KModule.h"
