+++ lib/Expr/Constraints.cpp
@@ -11,8 +11,8 @@
 
 #include "klee/Expr/ExprPPrinter.h"
 #include "klee/Expr/ExprVisitor.h"
+#include "klee/Support/OptionCategories.h"
 #include "klee/Module/KModule.h"
-#include "klee/OptionCategories.h"
 
 #include "llvm/IR/Function.h"
 #include "llvm/Support/CommandLine.h"
