+++ lib/Solver/ConstructSolverChain.cpp
@@ -11,10 +11,10 @@
  * This file groups declarations that are common to both KLEE and Kleaver.
  */
 
-#include "klee/Common.h"
+#include "klee/Solver/Common.h"
+#include "klee/Solver/SolverCmdLine.h"
 #include "klee/Support/ErrorHandling.h"
 #include "klee/System/Time.h"
-#include "klee/Solver/SolverCmdLine.h"
 
 #include "llvm/Support/raw_ostream.h"
