+++ lib/Solver/QueryLoggingSolver.cpp
@@ -9,11 +9,11 @@
 #include "QueryLoggingSolver.h"
 
 #include "klee/Config/config.h"
+#include "klee/Support/OptionCategories.h"
+#include "klee/Statistics/Statistics.h"
 #include "klee/Support/ErrorHandling.h"
 #include "klee/Support/FileHandling.h"
 #include "klee/System/Time.h"
-#include "klee/OptionCategories.h"
-#include "klee/Statistics.h"
 
 namespace {
 llvm::cl::opt<bool> DumpPartialQueryiesEarly(
